requirements.txt
numpy>=1.22.0
opencv-python>=4.5.5
torch>=1.12.0
torchvision>=0.13.0
ultralytics>=8.0.0  # YOLOv8 için
matplotlib>=3.5.0
scikit-learn>=1.0.0
scipy>=1.7.0
pillow>=9.0.0
pandas>=1.4.0
tqdm>=4.62.0
requests>=2.27.0
seaborn>=0.11.0
tensorboard>=2.8.0
PyYAML>=6.0
flask>=2.0.0  # Web arayüzü için (isteğe bağlı) 